requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Data analysis and technical indicators
pandas>=2.2.0
numpy>=1.26.0
alpha-vantage>=3.0.0

# Trading API
alpaca-trade-api>=3.0.0

# AI and OpenAI
openai>=1.35.0

# News API
newsapi-python>=0.2.6

# Scheduling and automation
schedule>=1.2.0

# Logging
structlog>=23.2.0

# Email (built-in modules, no need to install)
# smtplib and email are part of Python standard library
